fixing all lowercase fixing fileerroe function fixing true false operations
