--- Coreutils/true/main.go
@@ -1,11 +1,0 @@
-package main
-
-import "fmt"
-
-func True() int {
-	return 0
-}
-
-func main() {
-	fmt.Println(True())
-}
