+++ coreutils/false/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"os"
+)
+
+func main() {
+	os.Exit(1)
+}
