+++ coreutils/head/input.txt
