+++ coreutils/input.txt
