+++ coreutils/wc/input.txt
